+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/generic_capsule/data/response/NearbyARCapsuleSummaryResponse.java
@@ -34,6 +34,7 @@ public record NearbyARCapsuleSummaryResponse(
     @Schema(description = "캡슐 타입")
     CapsuleType capsuleType
 ) {
+
     public NearbyARCapsuleSummaryResponse {
         if (dueDate != null) {
             dueDate = dueDate.withZoneSameInstant(ResponseMappingConstant.ZONE_ID);
