+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group/data/dto/GroupDetailDto.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.function.Function;
 import site.timecapsulearchive.core.domain.group.data.response.GroupDetailResponse;
 import site.timecapsulearchive.core.domain.group.data.response.GroupMemberResponse;
-import site.timecapsulearchive.core.domain.group.data.response.GroupMemberSummaryResponse;
 
 public record GroupDetailDto(
     String groupName,
