--- backend/core/src/main/java/site/timecapsulearchive/core/domain/group/repository/GroupQueryRepository.java
@@ -1,34 +1,0 @@
-package site.timecapsulearchive.core.domain.group.repository;
-
-import static com.querydsl.core.group.GroupBy.groupBy;
-import static com.querydsl.core.group.GroupBy.list;
-import static site.timecapsulearchive.core.domain.group.entity.QGroup.group;
-import static site.timecapsulearchive.core.domain.group.entity.QMemberGroup.memberGroup;
-import static site.timecapsulearchive.core.domain.member.entity.QMember.member;
-
-import com.querydsl.core.types.Projections;
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.time.ZonedDateTime;
-import java.util.List;
-import java.util.Optional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
-import org.springframework.data.domain.SliceImpl;
-import org.springframework.stereotype.Repository;
-import site.timecapsulearchive.core.domain.group.data.dto.GroupDetailDto;
-import site.timecapsulearchive.core.domain.group.data.dto.GroupMemberDto;
-import site.timecapsulearchive.core.domain.group.data.dto.GroupSummaryDto;
-
-
-public interface GroupQueryRepository {
-
-    Slice<GroupSummaryDto> findGroupsSlice(
-        final Long memberId,
-        final int size,
-        final ZonedDateTime createdAt
-    );
-
-     Optional<GroupDetailDto> findGroupDetailByGroupId(final Long groupId);
-
-}
