+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group/service/GroupService.java
@@ -1,5 +1,8 @@
 package site.timecapsulearchive.core.domain.group.service;
 
+import site.timecapsulearchive.core.domain.group.service.read.GroupReadService;
+import site.timecapsulearchive.core.domain.group.service.write.GroupWriteService;
+
 public interface GroupService extends GroupReadService, GroupWriteService {
 
 }
