+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/group/service/GroupServiceImpl.java
@@ -8,6 +8,8 @@ import site.timecapsulearchive.core.domain.group.data.dto.GroupCreateDto;
 import site.timecapsulearchive.core.domain.group.data.dto.GroupDetailDto;
 import site.timecapsulearchive.core.domain.group.data.dto.GroupSummaryDto;
 import site.timecapsulearchive.core.domain.group.entity.Group;
+import site.timecapsulearchive.core.domain.group.service.read.GroupReadService;
+import site.timecapsulearchive.core.domain.group.service.write.GroupWriteService;
 
 @Service
 @RequiredArgsConstructor
